Add validation script and grunt task
Trying this with the eslint packages as direct dependencies, though
I’m not sure this will work. We may need to move them to be peer
dependencies.

Right now we don’t have any tests configured, but we can use a
circular dependency with the brinkbit-test module which has yet to be
published.